// frontend/src/Interfaces/BordingInterface.ts
export interface IbordingProcess {
  basicEmployeeDetail: {
    isActive: boolean;
    hasAccess: boolean;
    isSubmited: boolean;
  };
  EmployeesKyc: {
    isActive: boolean;
    hasAccess: boolean;
    isSubmited: boolean;
  };
  CabRequest: {
    isActive: boolean;
    hasAccess: boolean;
    isSubmited: boolean;
  };
}